Splitted module
