+++ .gitignore
@@ -1,2 +1,3 @@
 .idea
 _*
+!__init__.py
